3 - CurrencyListRaw moved to CurrencyRaw file
